DO NOT MERGE Remove cargo compat and nix tests
The modern updater tool adds two new fields to cargo imported packages.
Tokio does not strictly need either, and the build system for mainline
doesn't know about them yet, so they need to be removed to take the
update.

The tests in the new version of tokio are run on AOSP, but require an
updated version of the nix crate. To avoid a backport chain, I've simply
disabled those tests here.

Test: m
Bug: 206680413
Change-Id: I3fcc129036c9f1e8d166e838f1915d4c046b98fe